refactor: extract widget receivers
